Created separated files for Actions, Constants, Reducers and Utils
