+++ app/src/routes/Events/modules/FilterConstants.js
@@ -1,0 +1,10 @@
+export const SET_FILTER_ACTIVE = "SET_FILTER_ACTIVE";
+
+export const SET_FILTER_LOCATION_RADIUS = "SET_FILTER_LOCATION_RADIUS";
+
+export const SET_FILTER_DATE = "SET_FILTER_DATE";
+export const SET_FILTER_DATE_FROM = "SET_FILTER_DATE_FROM";
+export const SET_FILTER_DATE_TO = "SET_FILTER_DATE_TO";
+
+export const SET_FILTER_TOURNAMENT_ACTIVE = "SET_FILTER_TOURNAMENT_ACTIVE";
+export const SET_FILTER_SPARING_ACTIVE = "SET_FILTER_SPARING_ACTIVE";
