+++ app/src/routes/Events/modules/FilterReducers.js
@@ -1,77 +1,13 @@
-export const SET_FILTER_ACTIVE = "SET_FILTER_ACTIVE";
-
-export const SET_FILTER_LOCATION_RADIUS = "SET_FILTER_LOCATION_RADIUS";
-
-export const SET_FILTER_DATE = "SET_FILTER_DATE";
-export const SET_FILTER_DATE_FROM = "SET_FILTER_DATE_FROM";
-export const SET_FILTER_DATE_TO = "SET_FILTER_DATE_TO";
-
-export const SET_FILTER_TOURNAMENT_ACTIVE = "SET_FILTER_TOURNAMENT_ACTIVE";
-export const SET_FILTER_SPARING_ACTIVE = "SET_FILTER_SPARING_ACTIVE";
-
-export const setFilterActive = (id, active) => ({
-    type: SET_FILTER_ACTIVE,
-    payload: {
-        id,
-        active
-    }
-});
-
-export const setFilterLocationRadius = payload => ({
-    type: SET_FILTER_LOCATION_RADIUS,
-    payload
-});
-
-export const setFilterTournamentActive = payload => ({
-    type: SET_FILTER_TOURNAMENT_ACTIVE,
-    payload
-});
-
-export const setFilterSparingActive = payload => ({
-    type: SET_FILTER_SPARING_ACTIVE,
-    payload
-});
-
-export const setFilterDateRange = (from, to) => {
-    let type;
-    let payload;
-
-    if (from === null) {
-        type = SET_FILTER_DATE_TO;
-        payload = to;
-    } else if (to === null) {
-        type = SET_FILTER_DATE_FROM;
-        payload = from;
-    } else {
-        type = SET_FILTER_DATE;
-        payload = {
-            from,
-            to
-        };
-    }
-
-    return {
-        type,
-        payload
-    };
-};
-
-export function getCurrentDate() {
-    const date = new Date();
-    let day = date.getDate();
-    let month = date.getMonth();
-    const year = date.getFullYear();
-
-    if (day < 10) {
-        day = `0${day}`;
-    }
-
-    if (month < 10) {
-        month = `0${month}`;
-    }
-
-    return `${day}-${month}-${year}`;
-}
+import {
+    SET_FILTER_ACTIVE,
+    SET_FILTER_LOCATION_RADIUS,
+    SET_FILTER_TOURNAMENT_ACTIVE,
+    SET_FILTER_SPARING_ACTIVE,
+    SET_FILTER_DATE,
+    SET_FILTER_DATE_FROM,
+    SET_FILTER_DATE_TO
+} from "./FilterConstants";
+import getCurrentDate from "./FilterUtils";
 
 export function locationReducer(state = {}, { type, payload }) {
     if (type === SET_FILTER_LOCATION_RADIUS) {
